fix: added fade in / fade out logic for raindrop fx

+++ src/Features/WetnessEffects.h
@@ -59,7 +59,7 @@ public:
 	struct alignas(16) PerPass
 	{
 		float Time;
-		uint Raining;
+		float Raining;
 		float Wetness;
 		float PuddleWetness;
 		DirectX::XMFLOAT3X4 DirectionalAmbientWS;
